requirements.txt
beautifulsoup4==4.13.4
ecdsa==0.19.1
greenlet==3.2.3
MarkupSafe==3.0.2
playwright==1.53.0
pyee==13.0.0
python-dotenv==1.1.1
python-http-client==3.3.7
sendgrid==6.12.4
six==1.17.0
soupsieve==2.7
tenacity==9.1.2
typing_extensions==4.14.1
Werkzeug==3.1.3
